+++ neu-fall22-pdp-hw4/src/portfolio/views/impl/InfoPageView.java
@@ -1,9 +1,6 @@
 package portfolio.views.impl;
 
 import java.io.PrintStream;
-
-import javax.swing.JPanel;
-
 import portfolio.models.entities.PortfolioWithValue;
 import portfolio.views.ViewAbs;
 
@@ -25,8 +22,8 @@ public class InfoPageView extends ViewAbs {
    * @param errorMessage       the error message we want to show to the user
    */
   public InfoPageView(PrintStream printStream, Double costOfBasis,
-                      PortfolioWithValue portfolioWithPrice,
-                      String errorMessage) {
+      PortfolioWithValue portfolioWithPrice,
+      String errorMessage) {
     super(printStream);
     this.portfolioWithPrice = portfolioWithPrice;
     this.errorMessage = errorMessage;
@@ -41,7 +38,7 @@ public class InfoPageView extends ViewAbs {
    * @param errorMessage       the error message we want to show to the user
    */
   public InfoPageView(PortfolioWithValue portfolioWithPrice, Double costOfBasis,
-                      String errorMessage) {
+      String errorMessage) {
     this.portfolioWithPrice = portfolioWithPrice;
     this.errorMessage = errorMessage;
     this.costOfBasis = costOfBasis;
@@ -61,36 +58,36 @@ public class InfoPageView extends ViewAbs {
     printStream.println("----------------------------------------------------------");
     if (portfolioWithPrice != null) {
       printStream.println(
-              "If stock price not found, the value with be N/A and "
-                      + "will not be include in the total value.");
+          "If stock price not found, the value with be N/A and "
+              + "will not be include in the total value.");
       printStream.println("");
       printStream.println("Portfolio composition and value as of: "
-              + portfolioWithPrice.getDate());
+          + portfolioWithPrice.getDate());
       printStream.println("+---------+---------------+--------------------+");
       printStream.println("|    Stock|  No. of shares|       Current value|");
       printStream.println("+---------+---------------+--------------------+");
       for (var entry : portfolioWithPrice.getValues()) {
         String symbol = entry.getSymbol();
-        int amount = entry.getAmount();
+        double amount = entry.getAmount();
         String value = entry.getValue() ==
-                null ? "N/A" : "$" + entry.getValue().toString();
-        printStream.printf("|%9s|%15d|%20s|%n", symbol, amount, value);
+            null ? "N/A" : "$" + entry.getValue().toString();
+        printStream.printf("|%9s|%15.2f|%20s|%n", symbol, amount, value);
       }
       printStream.println("+---------+---------------+--------------------+");
       printStream.println("Total value: " + "$" + portfolioWithPrice.getTotalValue());
       if (costOfBasis != null) {
         printStream.println("Cost of basis as of "
-                + portfolioWithPrice.getDate() + ": " + "$" + costOfBasis);
+            + portfolioWithPrice.getDate() + ": " + "$" + costOfBasis);
       } else {
         printStream.println("Cost of basis is not available for this portfolio.");
       }
       printStream.println();
       printStream.println("Please enter the date again if "
-              + "you want to determine value for another date. "
-              + "The format is year-month-day, ex: 2022-10-11");
+          + "you want to determine value for another date. "
+          + "The format is year-month-day, ex: 2022-10-11");
     } else {
       printStream.println("Please enter the date that you want to determine. "
-              + "The format is year-month-day, ex: 2022-10-11");
+          + "The format is year-month-day, ex: 2022-10-11");
     }
     printStream.print("input > ");
   }
